--- Cinemate.Core/Entities/ApplicationRole.cs
@@ -1,15 +1,0 @@
-﻿using Microsoft.AspNetCore.Identity;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Cinemate.Core.Entities
-{
-	public class ApplicationRole : IdentityRole
-	{
-		public bool IsDefault { get; set; }
-		public bool IsDeleted { get; set; }
-	}
-}
