+++ Cinemate.Core/Entities/Auth/AuditableEntity.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Cinemate.Core.Entities.Auth
+{
+    public class AuditableEntity
+    {
+        public string CreatedById { get; set; } = string.Empty;
+        public DateTime CreatedOn { get; set; } = DateTime.UtcNow;
+        public string? UpdatedById { get; set; }
+        public DateTime? UpdatedOn { get; set; }
+        public ApplicationUser CreatedBy { get; set; } = default!;
+        public ApplicationUser? UpdatedBy { get; set; }
+    }
+}
