+++ Cinemate.Core/Errors/Authentication/AuthenticationError.cs
@@ -20,3 +20,4 @@ namespace Cinemate.Repository.Errors.Authentication
 		}
 	}
 }
+
