+++ Cinemate.Repository/Authentication/JwtProvider.cs
@@ -1,5 +1,5 @@
 ﻿using Cinemate.Core.Authentication_Contract;
-using Cinemate.Core.Entities;
+using Cinemate.Core.Entities.Auth;
 using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
 using System;
@@ -13,7 +13,7 @@ using System.Threading.Tasks;
 
 namespace Cinemate.Repository.Authentication
 {
-	public class JwtProvider : IJwtProvider
+    public class JwtProvider : IJwtProvider
 	{
 		private readonly JwtOptions _jwtOptions;
