+++ Cinemate.Repository/Data/Contexts/ApplicationDbContext.cs
@@ -1,4 +1,4 @@
-﻿using Cinemate.Core.Entities;
+﻿using Cinemate.Core.Entities.Auth;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
@@ -7,7 +7,7 @@ using System.Security.Claims;
 
 namespace Cinemate.Repository.Data.Contexts
 {
-	public class ApplicationDbContext : IdentityDbContext<ApplicationUser, ApplicationRole, string>
+    public class ApplicationDbContext : IdentityDbContext<ApplicationUser, ApplicationRole, string>
 	{
 		private readonly IHttpContextAccessor _httpContextAccessor;
 		public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options, IHttpContextAccessor httpContextAccessor) : base(options)
